Add reducer function to metrics (#133)
Summary:
Pull Request resolved: https://github.com/facebookincubator/dynolog/pull/133

## About
We can define hbt metrics in terms of a multiple events. However, we lack a way to specify the formula to combine these events,
today that is done manually outside in code or by the user directly.
We add an optional "reducer" functor to our metric definition that helps identify how to reduce performance events..

Included
* Reducer functor type definition and adding it to Metric class (Metric.h/.cpp.
* Perf ReadValues leverages the reducer functor to combine multiple events (CpuEventsGroup.h).
* Include a simple addition and rate reducer (BuiltinMetrics.h).
* Use the rate reducer in instructions/cycle.

Reviewed By: haowangludx

Differential Revision: D45092289

fbshipit-source-id: 2e044915d6595820eddbf95ed7dacfc73c38abfd