Update pytorch_profiler.md to include information about older PyTorch versions (#153)
Summary:
- Updates `pytorch_profiler.md` to include information about older PyTorch versions.

Based on [this thread](https://github.com/pytorch/kineto/issues/770), versions below PyTorch 2.0 do not work with the on-demand remote tracing feature in Dynolog.

It would be fruitful to mention this in the README  - currently, we only mention that the latest PyTorch 2.0 release is recommended, and nothing about the older versions.

Since using Dynolog with an older PyTorch version does not result in any kind of error or warning, it may be confusing to the users as to why they don't see the expected functionality. Thus, we can include this in the README since many still use older torch versions.

Pull Request resolved: https://github.com/facebookincubator/dynolog/pull/153

Reviewed By: sanrise

Differential Revision: D47109756

Pulled By: briancoutinho

fbshipit-source-id: 86d7dedca7cc5968874e95200463783c2425f0af