Replace memcmp with checked_memcmp (#177)
Summary:

No behavior change expected other than: A out-of-bound buffer access is theoretically possible. Looking at the code this is not a possibility, but a future change in the type names, or the type buffer size could have exposed an issue.
Marking the code as safe as is, as no inputs are caller/user controlled.

Reviewed By: briancoutinho

Differential Revision: D44138204