Replace memcmp with checked_memcmp (#177)
Summary:
Pull Request resolved: https://github.com/facebookincubator/dynolog/pull/177

No behavior change expected other than: A out-of-bound buffer access is theoretically possible. Looking at the code this is not a possibility, but a future change in the type names, or the type buffer size could have exposed an issue.
Marking the code as safe as is, as no inputs are caller/user controlled.

Reviewed By: briancoutinho

Differential Revision: D44138204

fbshipit-source-id: bb20ba529551afc750a51a00903d94a84e18586a