Remove extraneous semicolon from hbt/src/tagstack/Slicer.h (#190)
Summary:
Pull Request resolved: https://github.com/facebookincubator/dynolog/pull/190

Extraneous semicolons are a code smell and can mask more serious problems. `-Wextra-semi-stmt` finds them.

This diff removes an extraneous semicolon or adjusts a macro so that a semicolon is required after the macro (making it look like a standard function).

This file is drawn from a heavy-hitting list, so fixing this problem will allow *many* other files to take advantage of the safety `-Wextra-semi-stmt` offers.

This should be a low-risk diff: if it compiles, it works.

Reviewed By: meyering

Differential Revision: D51631119

fbshipit-source-id: 314b8c4dd22bc161cd390eff06339d1ac7436328