Remove unused variables in hbt/src/perf_event/BPerfEventsGroup.cpp (#211)
Summary:
Pull Request resolved: https://github.com/facebookincubator/dynolog/pull/211

LLVM-15 has a warning `-Wunused-but-set-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code, or (b) qualifies the variable with `[[maybe_unused]]`, mostly in cases where the variable _is_ used, but, eg, in an `assert` statement that isn't present in production code.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: danzimm, meyering

Differential Revision: D52848001

fbshipit-source-id: 6216c5a9143620394d61ac170fd0cb8f322f645c