Remove unused variables in dynolog/src/ScubaLogger.cpp (#213)
Summary:
Pull Request resolved: https://github.com/facebookincubator/dynolog/pull/213

LLVM-15 has a warning `-Wunused-but-set-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code, or (b) qualifies the variable with `[[maybe_unused]]`, mostly in cases where the variable _is_ used, but, eg, in an `assert` statement that isn't present in production code.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: danzimm, meyering

Differential Revision: D52848049

fbshipit-source-id: d1d62f0c35d864f7209a3c42b7bb94344250366a