Remove unused variables in hbt/src/perf_event/PmuDevices.cpp (#218)
Summary:
Pull Request resolved: https://github.com/facebookincubator/dynolog/pull/218

LLVM-15 has a warning `-Wunused-but-set-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code, or (b) qualifies the variable with `[[maybe_unused]]`, mostly in cases where the variable _is_ used, but, eg, in an `assert` statement that isn't present in production code.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: bunnypak, dmm-fb

Differential Revision: D53011652

fbshipit-source-id: d4acfb4694e8dde5bfebc56e7e21583155509ee4