make perf event value reset configurable from Monitor/PerCoreCountReader/PerUncoreCountreader (#256)
Summary:
Pull Request resolved: https://github.com/facebookincubator/dynolog/pull/256

PerfEventsGroup::enable() function takes a parameter `reset` that controls if it will PERF_EVENT_IOC_RESET its managed event before enable(). This is default to be true.

This is not the desired behavior we want for monitoring core/uncore perf events in dynolog. Since we are rarely directly work with PerfEventsGroup class but typically work with Monitor class, forward reset parameter up to Monitor so we can set this false when not needed.

Reviewed By: bigzachattack

Differential Revision: D56658256

fbshipit-source-id: b00dbdd6131ea342f9a7bf08d7530c98bce28127