Remove unused variables in gloo2/test/alltoallv_cpu_test.cc (#260)
Summary:
Pull Request resolved: https://github.com/facebookincubator/dynolog/pull/260

LLVM-15 has a warning `-Wunused-but-set-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code, or (b) qualifies the variable with `[[maybe_unused]]`, mostly in cases where the variable _is_ used, but, eg, in an `assert` statement that isn't present in production code.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Differential Revision: D57343994

fbshipit-source-id: 85a5f431633ffd5c3b77a7bea3945451daf5c0d8