Fix build with libc++
Previously dynolog relied on a fact that high_resolution_clock is an alias of system_clock, however in general it is not true[1].
On systems where high_resolution_clock is an alias of steady_clock no conversion is possible.
Instead of relying that high_resolution_clock is system_clock, it is better to use `system_clock::now` directly.

[1] https://en.cppreference.com/w/cpp/chrono/high_resolution_clock

Closes: #275