Fix build with libc++ (#276)
Summary:
Previously dynolog relied on a fact that `high_resolution_clock` is an alias of `system_clock`, however in general it is not true[1].
On systems where `high_resolution_clock` is an alias of `steady_clock` no conversion is possible.
Instead of relying that `high_resolution_clock` is an alias of `system_clock`, it is better to use `system_clock::now` directly.

[1] https://en.cppreference.com/w/cpp/chrono/high_resolution_clock

Closes: https://github.com/facebookincubator/dynolog/issues/275

Pull Request resolved: https://github.com/facebookincubator/dynolog/pull/276

Differential Revision: D61346530

Pulled By: briancoutinho

fbshipit-source-id: 745bfd3a45362e9d75afea492fa93759e7c2d184