hbt/bperf: Per thread monitoring core logic part 1 (#283)
Summary:
Pull Request resolved: https://github.com/facebookincubator/dynolog/pull/283

This is most of the non-perf_event part of the bperf per thread
monitoring:

1. Mechanism to share data between lead (e.g. dynolog) and per thread
   users (e.g. services). This is done via two pinned bpf map.
2. An extendable data format (struct bperf_thread_data) between lead
   and users. This is designed to provide some compatibility between
   dynolog and services with different versions of bperf.
3. Cumulative CPU time (cpuTime) and MONOTONIC clock (monoTime), for
   x86_64. Both of these require proper handling off TSC (or equivalent
   on arm64).
4. Basic userspace library BPerfPerThreadReader.

The code is designed to share with regular bperf. If the flag
support_per_thread is not set in the constructor, some maps in the
BPF skeleton will be loaded initially, but freed soon. This is achieved
by holding extra fd on the maps and programs in use, and destroying the
initial skeleton.

Reviewed By: Alston-Tang

Differential Revision: D61421284

fbshipit-source-id: 0badd2855a1d4e60b6c644a0b7d6def61a3a6317