do not enable/disble underlying bperf event in BPerfCountReader (#292)
Summary:
Pull Request resolved: https://github.com/facebookincubator/dynolog/pull/292

we will need to create BPerfCountReader upon global bperf event that is always on

so the expected behavior of BPerfCountReader is to just remove itself from the bpf map but do not change the underlying bperf event because some other BPerfCountReader instance may still use that event.

this should not change the behavior of current BPerf usage in TwTaskMonitor as we only create BPerfCountReader in cgroup mode.

Reviewed By: liu-song-6

Differential Revision: D61893592

fbshipit-source-id: 8d1c7e378233a456989f21b979f7c07ffdaabf81