hbt/bperf: Let per thread reader handle lead exit event. (#317)
Summary:
Pull Request resolved: https://github.com/facebookincubator/dynolog/pull/317

When the leader process exits, the per thread readers will not be able to
read data. This is by design. The readers need to detect such event and
renable the reader after the leader process comes back.

An enabled_ flag is added to the reader. We need a heuristic algorithm to
detect when the leader exits. Currently, the algorithm is simple: if we
read same value for the first hardware counter twice, the leader is most
likely gone. We may need to revisit this later.

A unit test is added to cover this. We also need to adjust close()
function of BPerfEventsGroup to make it easier for the reader to detect
leader exit events.

Reviewed By: Alston-Tang

Differential Revision: D64486001

fbshipit-source-id: 1226e2109526e329cd14561b4faa1fb541181b75