Move DynoTicker and DynoMonitor out of dyno/experimental (#346)
Summary:
Pull Request resolved: https://github.com/facebookincubator/dynolog/pull/346

DynoTicker is a simple class to schedule tick events on subscriber functions. Multiple levels of tick are supported, 'major', 'minor' and a templated number of levels of 'sub-minor' ticks. A typical cadence in telemetry could be 60s and 1s, for instrumenting course and fine grained.
Sub-minor ticks can be configured with custom sampling rates, which may be helpful if their work is too expensive to perform on every interval.
DynoMonitorBase is intended as an abstract base that uses the DynoTicker for scheduling.

Reviewed By: bigzachattack

Differential Revision: D69408689

fbshipit-source-id: be85c85bd4d5fa11859667de875502349da4c924