increase max allowed concurrent readers (#350)
Summary:
Pull Request resolved: https://github.com/facebookincubator/dynolog/pull/350

increase BPERF_MAX_THREAD_READER to 65535, also change the way how idx is allocated. Instead of using a bitmask, use a stack like structure to assign the idx on top. This is going to achieve a more efficient idx assignment on high concurrent # of readers, with a trade off of using 16x memory.

Reviewed By: liu-song-6

Differential Revision: D70737037

fbshipit-source-id: 0070e999dc58fafff56e619623bec7360a114ebb