allow 64k page sizes for arm
Summary:
Allow 64k page sizes for arm.

This code is used only for intel pt. I am going to used it for cs_etm on arm. Thus, page size is not limited by 4k and might be 64k too.

Checked the code around page_size and mmap_size variables. Followed all diffs until the diff where page_size was introduced. It seems that modifications to this check is safe.

Differential Revision: D71890002