allow 64k page sizes for arm (#355)
Summary:
Pull Request resolved: https://github.com/facebookincubator/dynolog/pull/355

Allow 64k page sizes for arm.

This code is used only for intel_pt. I am going to use it for cs_etm on arm too. Thus, page size is not limited by 4k and might be 64k.

Checked the code around page_size and mmap_size variables. Followed all diffs until the diff where page_size was introduced. It seems that modifications to this check is safe.

Reviewed By: jwiepert

Differential Revision: D71890002

fbshipit-source-id: c77a8c157d8a396fefbf8fea6d9daaae726ddc30