Make Prometheus HTTP server address configurable
Summary:
We are planning to deploy `dynolog` as a container in cloud.

To port forward the Prometheus metrics endpoint in a container to the host, we will need to configure it to listen to more IP addresses than just localhost. 

Implements a new flag that controls which address the Prometheus HTTP server listens on.

Differential Revision: D73493091