Make Prometheus HTTP server address configurable (#364)
Summary:
Pull Request resolved: https://github.com/facebookincubator/dynolog/pull/364

We are planning to deploy `dynolog` as a container in cloud.

To port forward the Prometheus metrics endpoint in a container to the host, we will need to configure it to listen to more IP addresses than just localhost.

Implements a new flag that controls which address the Prometheus HTTP server listens on.

Reviewed By: jj10306

Differential Revision: D73493091

fbshipit-source-id: f88b3dbf53cec8816cfc32845d520f869fb753bc