fix incosistent enalbe/disable status during the commitElemChangeOnLocalCpu_ (#365)
Summary:
Pull Request resolved: https://github.com/facebookincubator/dynolog/pull/365

we cannot use isEnabled() in commitElemChangeOnLocalCpu_. isEnabled() assume that all of the readers in the same CountReader should have consistent enable/disable/open/close status. however, this is no longer true in commitElemChangeOnLocalCpu_() because it will enable a portion of readers belonged to the same CPU first

always use is(Enable||Open)OnCpu to make sure we only check the status consistency for readers opened for the same CPU.

since commitElemChangeOnLocalCpu_() is not a public interface, we could limit the change within this function and don't need to worry about potential misuse from a hbt user.

Reviewed By: bigzachattack

Differential Revision: D73643269

fbshipit-source-id: 56b5d27ae31b2fc181e27afd8a89c75df2f07b5e