Fix Kineto OSS Logger (#385)
Summary:
Pull Request resolved: https://github.com/facebookincubator/dynolog/pull/385

In D56577485, we removed the Kineto case for logging from the FabricManager in OSS for some reason. Because of this we see build failures in this PR: https://github.com/pytorch/kineto/actions/runs/15552720676/job/43832280645?pr=1101

We need to reintroduce it so we can bump up the version again.

Reviewed By: dmm-fb

Differential Revision: D76372804

fbshipit-source-id: eb7b42b7a56dd36cbfab686512603ab4468bf22c