Fix `/proc/interrupts` parsing on Bergamos (#386)
Summary:
Pull Request resolved: https://github.com/facebookincubator/dynolog/pull/386

Bergamos have a lot of cores and `/proc/interrupts` have entry for each core.
`InterruptStatsMonitor` was parsing `/proc/interrupts` into `buf[1024]`, which
is not enough to accomodate one line. For example, `TLB shootdowns` line from
the random box is 1957 bytes.

Entire line doesn't fit into `buf` and `getline` splits in few lines and then
all the logic stopped working: we can not match `TLB shootdowns` substring
anymore and `valueCount` doesn't match with `cpuCount_` and as a results we
exported metric is invalid.

Reviewed By: ot

Differential Revision: D76141968

fbshipit-source-id: b0d6c12887c920eba93cc4bcd35ff9aa953264fe