Use cpu core count in SystemInfo for cpu utilization (#439)
Summary:

In a number of places, we use `std::thread::hardware_concurrency()` expecting it to return the number of available cpu threads of the entire host. This assumes the cgroup that Dynolog runs in has no restrictions on the cpuset. Today we discovered this is not true:

```
[root@twstorage15779.pci1 ~]# cat /sys/fs/cgroup/cpuset.cpus.effective
0-51
[root@twstorage15779.pci1 ~]# cat /sys/fs/cgroup/system.slice/cpuset.cpus
0-5,26-31
```

This caused `std::thread::hardware_concurrency()` on this host to evaluate to 12 instead of 52, resulting in both incorrect cpu usage and cpu utilization % being reported. This diff fixes the issue by replacing `std::thread::hardware_concurrency()` with `systemInfo->cpuCountInt`. `systemInfo->cpuCountInt` reads from dynowhatami first and falls back to `sysconf(_SC_NPROCESSORS_CONF)` if CpuCount is invalid in dynowhatami. It should be a safe alternative.

Reviewed By: cmgrace, abhiShedge

Differential Revision: D81073488