Apply LICENSELINT to OSS Dynolog (#443)
Summary:
Pull Request resolved: https://github.com/facebookincubator/dynolog/pull/443

We received an alert that we are failing to meeting OSS requirements due to missing or incorrect headers. I found out that we actually have LICENSELINT (https://www.internalfb.com/wiki/Linting/License_Lint/) that can automatically fix the headers and should help catching this issue at diff time in the future.

Reviewed By: cmgrace

Differential Revision: D81940929