Add Per CpuSocket scope and reading per CpuSocket uncore metric support to hbt (#458)
Summary:

This diff introduce the CpuSocket scope to uncore_scope in hbt library and added socket memory latency metric definitions to ArmUncoreMetrics.

The apis that expect an uncore Scope now takes a vector of uncore scope so that the cpu to scope mapping can be tracked internally. Value reading uses the same logic as
`readAllUncoreCountsPerPerfEventOnLocalCpu`, where dynolog is context switched to the same cpu as uncore counter to read metric without causing IPIs.

The assumption here is that
1. CPU0 maps to socket 0 and
2. socket 0 correspond to the perf counters nvidia_scf_pmu/socket_0_*

Reviewed By: Alston-Tang

Differential Revision: D84030235