Modifying the flow to capture the IPC channel
Summary:
We added a few modifications to the code where the registration code for ondemand can idenitify the IPC channel that the Kineto process used. This information is then added to profiler results to provide clients like Dyno CLI a good signal to indicate if a given trace generated by on-demand tracing will have been uploaded to Manifold.

OSS pytorch running on Mast like Conda on Mast will have this set to false indicating that the kineto runtime in those workloads does not have the manifold driver to upload files to object storage.

Differential Revision: D86722300